refactor: extract resolveAgentTools into dedicated runtime module
The tool resolution pipeline — disabled filter, unavailable filter,
delegation gate, and conditional injections — was implemented as a
private function inside runToolUseFlow.ts.  Any developer asking
"what tools does an agent run with?" had to trace into a flow
implementation rather than finding a named module.

Move the logic to src/agent/runtime/agentToolResolution.ts with a
doc-comment that spells out the five-step pipeline.  runToolUseFlow.ts
now imports a single resolveAgentTools() call instead of pulling from
four separate @tools/ and @agent/runtime/ modules directly.

No behaviour change.
